Added checks and modularity

+++ install.sh
@@ -2,8 +2,11 @@
 echo "This could take a while... Go sleep\n"
 
 # Delete link to examples. Because.
-echo "Deleting link to examples"
-rm ~/examples.desktop
+if [ -e "~/examples.desktop" ]
+then
+	echo "Deleting link to examples"
+	rm ~/examples.desktop
+fi
 
 # Add ppas
 echo "Adding personal package archives"
@@ -33,12 +36,42 @@ apt-get clean
 
 # Download programs from sources
 echo "Getting newer versions...\n"
-if [ ! -e "eclipse-SDK-4.2.2-linux-gtk-x86_64.tar.gz" ]; then
-	wget "http://ftp.osuosl.org/pub/eclipse/eclipse/downloads/drops4/R-4.2.2-201302041200/eclipse-SDK-4.2.2-linux-gtk-x86_64.tar.gz"
+ECLVER="eclipse-SDK-4.2.2-linux-gtk-x86_64.tar.gz"
+if [ ! -e $ECLVER ]; then
+	ECLDIR="R-4.2.2-201302041200/"
+	
+	echo "Dowloading $ECLVER"
+	wget "http://ftp.osuosl.org/pub/eclipse/eclipse/downloads/drops4/$ECLDIR$ECLVER"
+	
+	echo "Unzipping $ECLVER"
+	tar -xzvf $ECLVER
+
+	echo "Removing archive"
+	rm $ECLVER
+
+	echo "Moving old version"
+	mv /usr/lib/eclipse /usr/lib/eclipse-old
+	
+	echo "Installing new eclipse"
+	mv eclipse /usr/lib/eclipse
 fi
 
-if [ ! -e "ideaIC-12.1.3.tar.gz" ]; then
-	wget "http://download-ln.jetbrains.com/idea/ideaIC-12.1.3.tar.gz"
+IDEAVER="ideaIC-12.1.4.tar.gz"
+if [ ! -e $IDEAVER ]; then
+	echo "Downloading $IDEAVER"
+	wget "http://download-ln.jetbrains.com/idea/$IDEAVER"
+	
+	echo "Unzipping $IDEAVER"
+        tar -xzvf $IDEAVER
+
+        echo "Removing archive"
+        rm $IDEAVER
+
+        echo "Moving old version"
+        mv /opt/intellij-idea-ce /opt/intellij-idea-ce-old
+
+        echo "Installing new IntelliJ IDEA"
+        # Haven't figured this out yet
 fi
 
 # Say goodbye
